Reference `shared` as a symlink
