+++ app/backend/src/config/config.local.ts
@@ -11,7 +11,7 @@ import {
 } from "./Config";
 import { Env } from "@/tools/Env";
 import { LogLevel } from "@/tools/Logger";
-import chains from "../../../shared/chains.json";
+import chains from "@/shared/chains.json";
 
 export function getLocalConfig(env: Env): Config {
   const databaseConfig: DatabaseConfig = {
