+++ app/backend/src/core/clients/coingecko/assets/addL2Addresses.ts
@@ -3,7 +3,7 @@ import path from "path";
 import axios from "axios";
 import { WhitelistedAsset } from "@/core/clients/coingecko/assets/types";
 import { Logger } from "@/tools/Logger";
-import chains from "../../../../../../shared/chains.json";
+import chains from "@/shared/chains.json";
 import { ethers } from "ethers";
 import { getChecksumAddress } from "starknet";
