+++ app/backend/src/core/controllers/appraiser/handlers/BlockRewardsHandler.ts
@@ -5,7 +5,7 @@ import {
   Block,
   TransactionReceipt,
 } from "@/core/clients/blockchain/IBlockchainClient";
-import chains from "../../../../../../shared/chains.json";
+import chains from "@/shared/chains.json";
 
 export interface BlockRewardSummary {
   gasFees: bigint;
