+++ app/backend/src/database/repositories/SyncStatusRepository.test.ts
@@ -2,11 +2,11 @@ import { Knex } from "knex";
 import { expect } from "earl";
 import {
   SyncStatusRecord,
-  SubmissionType,
   SyncStatusRepository,
   TABLE_NAME,
 } from "./SyncStatusRepository";
 import { getTestDatabase } from "@/database/getTestDatabase";
+import { SubmissionType } from "@/shared/api/viewModels/SyncStatusEndpoint";
 
 describe(SyncStatusRepository.name, () => {
   let repository: SyncStatusRepository;
