+++ app/backend/src/database/repositories/SyncStatusRepository.ts
@@ -1,15 +1,9 @@
 import { Knex } from "knex";
 import { UnixTime } from "@/core/types/UnixTime";
 import { chainTableMapping } from "@/database/repositories/BlockValueRepository";
+import { SubmissionType } from "@/shared/api/viewModels/SyncStatusEndpoint";
 
 export const TABLE_NAME = "sync_status";
-
-export enum SubmissionType {
-  DataSubmission = "data_submission",
-  L2Finalization = "l2_finalization",
-  ProofSubmission = "proof_submission",
-  StateUpdates = "state_updates",
-}
 
 export interface SyncStatusRecord {
   chain_id: number;
