+++ app/frontend/src/pages/index.tsx
@@ -3,15 +3,15 @@ import { Box, Divider, Grid, Paper, Typography } from '@mui/material'
 import '@/app/globals.css'
 import RootLayout from '@/components/RootLayout'
 import ChainSelector from '@/components/ChainSelector'
-import { HealthStatusViewModel } from '../../../shared/api/viewModels/HealthEndpoint'
+import { HealthStatusViewModel } from '@/shared/api/viewModels/HealthEndpoint'
 import { healthApi } from '@/api/healthApi'
 import { metadataApi } from '@/api/metadataApi'
 import { syncStatusApi } from '@/api/syncStatusApi'
-import { MetadataRecordViewModel } from '../../../shared/api/viewModels/MetadataEndpoint'
+import { MetadataRecordViewModel } from '@/shared/api/viewModels/MetadataEndpoint'
 import {
   SyncStatusViewModel,
   VaRLiveDataViewModel,
-} from '../../../shared/api/viewModels/SyncStatusEndpoint'
+} from '@/shared/api/viewModels/SyncStatusEndpoint'
 import VaRLiveSection from './VaRLiveSection'
 import VaRHistoryChart from '@/components/charts/VaRHistoryChart'
 import SyncStatusTable from '@/components/charts/SyncStatusTable'
@@ -19,7 +19,7 @@ import moment from 'moment'
 import { dropdownCardPaperStyle } from '@/styles/dropdownCardStyles'
 import FinalityTimeseries from '@/components/charts/FinalityTimeseries'
 import { blocksApi } from '@/api/blocksApi'
-import { LatestBlockViewModel } from '../../../shared/api/viewModels/BlocksEndpoint'
+import { LatestBlockViewModel } from '@/shared/api/viewModels/BlocksEndpoint'
 import IndexerStatus from '@/components/ui/IndexerStatus'
 
 const pageTitle: string = 'L2 Finality Dashboard'
